push my "virtual mailman patch" supported in alternc
this changes the way lists are created and stored in the alternc
database to allow lists with similar names to exist on different domains
(ever wanted info@example.com and info@exemple.com be different lists?
this is it.) the stored "name" is foo-bar where the list address is
foo@bar.com

this patch therefore removes the "one list name per site" limitation
from alternc itself, but you need to patch mailman itself for it to
support that functionality. 

see http://wiki.koumbit.net/VirtualMailman

this should work for both unpatched and patched mailman installations